norminette

+++ src/extra/ft_nbr_len_unsigned.c
@@ -6,13 +6,13 @@
 /*   By: skoulen <skoulen@student.42lausanne.ch>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/06 13:15:53 by skoulen           #+#    #+#             */
-/*   Updated: 2022/11/06 13:18:09 by skoulen          ###   ########.fr       */
+/*   Updated: 2022/11/07 10:52:44 by skoulen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_nbr_len_unsigned(unsigned int n, unsigned int base)
+int	ft_nbr_len_unsigned(size_t n, size_t base)
 {
 	int	len;
